# shared/docs/api-spec.yaml
openapi: 3.1.0
info:
  title: CreditCatch API
  version: 1.0.0
  description: |
    API for CreditCatch — track and maximize credit card benefits.
    All endpoints except auth require a Bearer JWT in the Authorization header.

servers:
  - url: http://localhost:8080/api/v1
    description: Local development

security:
  - bearerAuth: []

tags:
  - name: auth
    description: Authentication and session management
  - name: cards
    description: Credit card catalog
  - name: user-cards
    description: User's card collection
  - name: credits
    description: Credit tracking and usage
  - name: dashboard
    description: Dashboard summaries

paths:
  # ── Auth ────────────────────────────────────────────

  /auth/signup:
    post:
      tags: [auth]
      summary: Create a new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"

  /auth/login:
    post:
      tags: [auth]
      summary: Log in with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/logout:
    post:
      tags: [auth]
      summary: Invalidate current session
      responses:
        "204":
          description: Logged out

  # ── Card Catalog ────────────────────────────────────

  /cards:
    get:
      tags: [cards]
      summary: List all credit cards in the catalog
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: issuer
          in: query
          schema:
            type: string
          description: Filter by card issuer
        - name: q
          in: query
          schema:
            type: string
          description: Search by card name or issuer
      responses:
        "200":
          description: Paginated list of cards
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Card"
                  next_cursor:
                    type: string
                    nullable: true

  /cards/{card_id}:
    get:
      tags: [cards]
      summary: Get card details including all credit definitions
      parameters:
        - $ref: "#/components/parameters/CardID"
      responses:
        "200":
          description: Card with credit definitions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── User Cards ──────────────────────────────────────

  /me/cards:
    get:
      tags: [user-cards]
      summary: List user's cards
      responses:
        "200":
          description: User's card collection
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserCard"

    post:
      tags: [user-cards]
      summary: Add a card to user's collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddUserCardRequest"
      responses:
        "201":
          description: Card added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCard"
        "409":
          description: User already has this card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"

  /me/cards/{user_card_id}:
    get:
      tags: [user-cards]
      summary: Get a user card with its credit periods
      parameters:
        - $ref: "#/components/parameters/UserCardID"
      responses:
        "200":
          description: User card detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCardDetail"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [user-cards]
      summary: Update a user card
      parameters:
        - $ref: "#/components/parameters/UserCardID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserCardRequest"
      responses:
        "200":
          description: Card updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCard"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [user-cards]
      summary: Remove a card from user's collection
      parameters:
        - $ref: "#/components/parameters/UserCardID"
      responses:
        "204":
          description: Card removed
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Credits ─────────────────────────────────────────

  /me/credits:
    get:
      tags: [credits]
      summary: List all credit periods for the user
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: status
          in: query
          schema:
            type: string
            enum: [used, unused, all]
            default: all
          description: Filter by usage status
        - name: user_card_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by specific user card
      responses:
        "200":
          description: Paginated credit periods
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CreditPeriod"
                  next_cursor:
                    type: string
                    nullable: true

  /me/credits/current:
    get:
      tags: [credits]
      summary: Get all active credit periods (current month/year)
      responses:
        "200":
          description: Current credit periods grouped by card
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CurrentCredit"

  /me/credits/{credit_period_id}/mark-used:
    post:
      tags: [credits]
      summary: Mark a credit as used
      parameters:
        - $ref: "#/components/parameters/CreditPeriodID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount_used_cents:
                  type: integer
                  description: Amount used in cents. Defaults to the full credit amount.
                transaction_id:
                  type: string
                  format: uuid
                  description: Optionally link to a transaction
      responses:
        "200":
          description: Credit marked as used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditPeriod"
        "404":
          $ref: "#/components/responses/NotFound"

  /me/credits/{credit_period_id}/mark-unused:
    post:
      tags: [credits]
      summary: Revert a credit to unused
      parameters:
        - $ref: "#/components/parameters/CreditPeriodID"
      responses:
        "200":
          description: Credit marked as unused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditPeriod"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Dashboard ───────────────────────────────────────

  /me/dashboard/summary:
    get:
      tags: [dashboard]
      summary: Dashboard summary for current period
      responses:
        "200":
          description: Summary of credit usage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /me/dashboard/annual:
    get:
      tags: [dashboard]
      summary: Annual credit usage breakdown
      parameters:
        - name: year
          in: query
          schema:
            type: integer
          description: Year to summarize (defaults to current year)
      responses:
        "200":
          description: Annual breakdown by card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnualSummary"

  /me/dashboard/monthly:
    get:
      tags: [dashboard]
      summary: Monthly credit usage breakdown
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        "200":
          description: Monthly breakdown by card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlySummary"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor for pagination
    CardID:
      name: card_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserCardID:
      name: user_card_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CreditPeriodID:
      name: credit_period_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    # ── Auth ────────────────────────────────────────

    SignupRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, email, name, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time

    # ── Card Catalog ──────────────────────────────

    Card:
      type: object
      required: [id, issuer, name, network, annual_fee, active]
      properties:
        id:
          type: string
          format: uuid
        issuer:
          type: string
        name:
          type: string
        network:
          type: string
          enum: [visa, mastercard, amex, discover]
        annual_fee:
          type: integer
          description: Annual fee in cents
        image_url:
          type: string
        active:
          type: boolean

    CardDetail:
      allOf:
        - $ref: "#/components/schemas/Card"
        - type: object
          required: [credits]
          properties:
            credits:
              type: array
              items:
                $ref: "#/components/schemas/CreditDefinition"

    CreditDefinition:
      type: object
      required: [id, name, amount_cents, period, category]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        amount_cents:
          type: integer
        period:
          type: string
          enum: [monthly, annual, one_time]
        category:
          type: string

    # ── User Cards ────────────────────────────────

    UserCard:
      type: object
      required: [id, card, nickname, active, created_at]
      properties:
        id:
          type: string
          format: uuid
        card:
          $ref: "#/components/schemas/Card"
        nickname:
          type: string
        opened_date:
          type: string
          format: date
          nullable: true
        annual_fee_date:
          type: string
          format: date
          nullable: true
        statement_close_day:
          type: integer
          minimum: 1
          maximum: 31
          nullable: true
          description: Day of month the statement closes (drives monthly credit periods)
        active:
          type: boolean
        created_at:
          type: string
          format: date-time

    UserCardDetail:
      allOf:
        - $ref: "#/components/schemas/UserCard"
        - type: object
          required: [credit_periods]
          properties:
            credit_periods:
              type: array
              items:
                $ref: "#/components/schemas/CreditPeriod"

    AddUserCardRequest:
      type: object
      required: [card_catalog_id, statement_close_day]
      properties:
        card_catalog_id:
          type: string
          format: uuid
        nickname:
          type: string
        opened_date:
          type: string
          format: date
        annual_fee_date:
          type: string
          format: date
        statement_close_day:
          type: integer
          minimum: 1
          maximum: 31
          description: Day of month the statement closes (required for monthly credit tracking)

    UpdateUserCardRequest:
      type: object
      properties:
        nickname:
          type: string
        opened_date:
          type: string
          format: date
          nullable: true
        annual_fee_date:
          type: string
          format: date
          nullable: true
        statement_close_day:
          type: integer
          minimum: 1
          maximum: 31
          nullable: true
        active:
          type: boolean

    # ── Credits ───────────────────────────────────

    CreditPeriod:
      type: object
      required: [id, user_card_id, credit_definition, period_start, period_end, used, amount_used_cents]
      properties:
        id:
          type: string
          format: uuid
        user_card_id:
          type: string
          format: uuid
        credit_definition:
          $ref: "#/components/schemas/CreditDefinition"
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        used:
          type: boolean
        used_at:
          type: string
          format: date-time
          nullable: true
        amount_used_cents:
          type: integer
        transaction_id:
          type: string
          format: uuid
          nullable: true

    CurrentCredit:
      type: object
      required: [user_card, credits]
      properties:
        user_card:
          $ref: "#/components/schemas/UserCard"
        credits:
          type: array
          items:
            $ref: "#/components/schemas/CreditPeriod"

    # ── Dashboard ─────────────────────────────────

    DashboardSummary:
      type: object
      required: [total_available_cents, total_used_cents, total_unused_cents, credits_used, credits_total, cards]
      properties:
        total_available_cents:
          type: integer
        total_used_cents:
          type: integer
        total_unused_cents:
          type: integer
        credits_used:
          type: integer
        credits_total:
          type: integer
        cards:
          type: array
          items:
            $ref: "#/components/schemas/CardSummary"

    CardSummary:
      type: object
      required: [user_card_id, card_name, issuer, available_cents, used_cents, credits_used, credits_total]
      properties:
        user_card_id:
          type: string
          format: uuid
        card_name:
          type: string
        issuer:
          type: string
        available_cents:
          type: integer
        used_cents:
          type: integer
        credits_used:
          type: integer
        credits_total:
          type: integer

    AnnualSummary:
      type: object
      required: [year, total_available_cents, total_used_cents, months]
      properties:
        year:
          type: integer
        total_available_cents:
          type: integer
        total_used_cents:
          type: integer
        months:
          type: array
          items:
            type: object
            required: [month, available_cents, used_cents]
            properties:
              month:
                type: integer
              available_cents:
                type: integer
              used_cents:
                type: integer

    MonthlySummary:
      type: object
      required: [year, month, total_available_cents, total_used_cents, cards]
      properties:
        year:
          type: integer
        month:
          type: integer
        total_available_cents:
          type: integer
        total_used_cents:
          type: integer
        cards:
          type: array
          items:
            $ref: "#/components/schemas/CardSummary"

    # ── Common ────────────────────────────────────

    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Field-level validation errors